+++ lib/generators/templates/knock.rb
@@ -3,7 +3,8 @@ Knock.setup do |config|
   ## Expiration claim
   ## ----------------
   ##
-  ## How long before a token is expired.
+  ## How long before a token is expired. If nil is provided, token will
+  ## last forever.
   ##
   ## Default:
   # config.token_lifetime = 1.day