+++ test/test_helper.rb
@@ -34,5 +34,6 @@ class ActiveSupport::TestCase
     Knock.token_secret_signature_key = -> { Rails.application.secrets.secret_key_base }
     Knock.token_public_key = nil
     Knock.token_audience = nil
+    Knock.token_lifetime = 1.day
   end
 end