+++ app/model/knock/auth_token.rb
@@ -7,11 +7,13 @@ module Knock
 
     def initialize payload: {}, token: nil
       if token.present?
-        @payload, _ = JWT.decode token, key, true, verify_claims
+        @payload, _ = JWT.decode token, decode_key, true, options
         @token = token
       else
         @payload = payload
-        @token = JWT.encode(claims.merge(payload), key, 'HS256')
+        @token = JWT.encode claims.merge(payload),
+          secret_key,
+          Knock.token_signature_algorithm
       end
     end
 
@@ -28,21 +30,49 @@ module Knock
     end
 
   private
-    def key
+    def secret_key
       Knock.token_secret_signature_key.call
     end
 
+    def decode_key
+      Knock.token_public_key || secret_key
+    end
+
+    def options
+      verify_claims.merge({
+        algorithm: Knock.token_signature_algorithm
+      })
+    end
+
     def claims
-      {
-        exp: Knock.token_lifetime.from_now.to_i,
-        aud: Knock.token_audience
-      }
+      _claims = {}
+      _claims[:exp] = token_lifetime if verify_lifetime?
+      _claims[:aud] = token_audience if verify_audience?
+      _claims
+    end
+
+    def token_lifetime
+      Knock.token_lifetime.from_now.to_i if verify_lifetime?
+    end
+
+    def verify_lifetime?
+      !Knock.token_lifetime.nil?
     end
 
     def verify_claims
       {
-        aud: Knock.token_audience, verify_claims: Knock.token_audience.present?
+        aud: token_audience,
+        verify_aud: verify_audience?,
+        verify_expiration: verify_lifetime?
       }
+    end
+
+    def token_audience
+      verify_audience? && Knock.token_audience.call
+    end
+
+    def verify_audience?
+      Knock.token_audience.present?
     end
   end
 end