+++ test/test_helper.rb
@@ -21,3 +21,19 @@ if ActiveSupport::TestCase.respond_to?(:fixture_path=)
   ActiveSupport::TestCase.fixture_path = File.expand_path("../fixtures", __FILE__)
   ActiveSupport::TestCase.fixtures :all
 end
+
+# Make sure knock global configuration is reset before every tests
+# to avoid order dependent failures.
+class ActiveSupport::TestCase
+  setup :reset_knock_configuration
+
+  private
+
+  def reset_knock_configuration
+    Knock.token_signature_algorithm = 'HS256'
+    Knock.token_secret_signature_key = -> { Rails.application.secrets.secret_key_base }
+    Knock.token_public_key = nil
+    Knock.token_audience = nil
+    Knock.token_lifetime = 1.day
+  end
+end