+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [2.0.0] - Unreleased
+## Added
+- Possibility to have permanent tokens
 
 ### Upgrade from 1.3